Récente avancé pour le traitement des images
images sont traité, affichées, peuvent être modifié/supprimé uniquement par le créateur

+++ back-end/controllers/sauce.js
@@ -1,24 +1,59 @@
 const Sauce = require('../models/Sauce');
+const fs = require('fs');
 
 exports.createSauce =  (req, res) => {
-  const sauce = new Sauce ({
-    ...req.body
-  });
-  sauce.save()
-    .then(() => res.status(201).json({ message: "Sauce bien ajoutée"}))
-    .catch(error => res.status(400).json({ error: error }));
+    const sauceObject = JSON.parse(req.body.sauce);
+    delete sauceObject._id;
+    delete sauceObject._userId;
+    const sauce = new Sauce({
+      ...sauceObject,
+      userId: req.auth.userId,
+      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    });
+    sauce.save()
+      .then(() => res.status(201).json({ message: "Sauce bien ajoutée"}))
+      .catch(error => res.status(400).json({ error: error }));
 };
 
 exports.modifySauce =  (req, res) => {
-  Sauce.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Sauce modifié'}))
-    .catch(error => res.status(400).json({ error: error }));
+  const sauceObject = req.file ? {
+    ...JSON.parse(req.body.sauce),
+    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+  } : { ...req.body };
+
+  delete sauceObject._userId;
+  Sauce.findOne({_id: req.params.id})
+    .then((sauce) => {
+      if (sauce.userId != req.auth.userId) {
+          res.status(401).json({ message : 'Not authorized'});
+      } else {
+          Sauce.updateOne({ _id: req.params.id}, { ...sauceObject, _id: req.params.id})
+          .then(() => res.status(200).json({message : 'Objet modifié!'}))
+          .catch(error => res.status(401).json({ error }));
+      }
+    })
+    .catch((error) => {
+      res.status(400).json({ error });
+    });
 };
 
-exports.deleteSauce =  (req, res) => {
-  Sauce.deleteOne({ _id: req.params.id })
-  .then(() => res.status(200).json({ message: 'Sauce supprimé'}))
-  .catch(error => res.status(400).json({ error: error }));
+exports.deleteSauce = (req, res, next) => {
+  Sauce.findOne({ _id: req.params.id})
+    .then(sauce => {
+        if (sauce.userId != req.auth.userId) {
+          res.status(401).json({message: 'Not authorized'});
+        } else {
+            const filename = sauce.imageUrl.split('/images/')[1];
+            fs.unlink(`images/${filename}`, () => {
+              Sauce.deleteOne({_id: req.params.id})
+                .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                .catch(error => res.status(401).json({ error }));
+            });
+        }
+    })
+    .catch( error => {
+      res.status(500).json({ error });
+    });
 };
 
 exports.displayOneSauce =  (req, res) => {
